configure.ac: require Mono 2.9
The README file was still kind of confusing in the Requirements section,
because:
a) It stated "mono trunk" instead of "mono master", as svn->git migration
happened already a while ago.
b) It mentioned a specific commit in mono instead of simply the lowest
Mono version that included this commit.

Looking at the commit, that fixes bug BNC#641146, it is dated on
September 2010 and went to the master branch:
https://bugzilla.novell.com/show_bug.cgi?id=641146

On August 2012, master branch gets labeled from "2.8.1" to "2.9":
http://lists.ximian.com/mailman/private/mono-patches/2010-October/178421.html

So we can safely assume that any Mono version equal or higher than
Mono 2.9 contains this fix. (The fix was backported too to the 2-8
branch but I still don't know which of the 2.8.x series included it;
at least I know it is not 2.8.1 as that is the version that the
reporter of the bug had.)