targets.make: fix generation of Microsoft.FSharp.targets
This file was ending up installed as is, without the substitution
required to adjust some values such as {BuildSuffix}. As a
consequence, the file could not be used and xbuild would give an
error when trying to find the dll specified.

The values to use forsubst.exe were taken from FSharpSource.targets
file, which has a <Target> to call it, and was not being used
because xbuild is not (and cannot be, I suppose) used when compiling
with Mono due to bootstrapping issues.