configure.ac: warn about not using the same prefix as mono
Lots of users think fsharp for Linux default installation path is
different than mono "by default". What happens is that the configuration
phase just configures the installation prefix as /usr/local by default
if nothing is supplied.

To better clarify this, we add a warning.
(As integration with xbuild is not likely going to work if different
prefix is used, as I haven't tested that scenario yet.)