Add policy files for FSharp.Core 2.0.0.0,4.0.0.0 --> 2.3.0.0, 4.3.0.0
F# 3.0 uses version number 4.3.0.0 for FSharp.Core (2.3.0.0 for the
Mono-for-.NET-2.0 profile). This would break binary compat for
assemblies that expect to bind to FSharp.Core 4.0.0.0.

This means we need to install a global policy redirect DLL for
FSharp.Core 2.0.0.0 --> FSharp.Core 2.3.0.0
FSharp.Core 4.0.0.0 --> FSharp.Core 4.3.0.0

For example, the existing MonoDevelop binding DLL for F# expects to
bind to FSharp.Core 4.0.0.0.

Also, "make do-4-0" no longer makes the proto compiler, you need a
"make all" or "make proto" for that