Optionally build FSharp.Core for Mono profile 2.1 (MonoAndroid, on Mac).
Optionally build FSharp.Core for Mono profile 2.1.

    make do-2-1

builds

   libs/debug/2.1/FSharp.Core

which (fingers crossed!) should be enough or MonoAndroid and MonoTouch,
though the restrictions on generics on those platforms will be a
problem.

Only works for Mac with MonoAndroid installed right now but only
because the path to the 2.1 profile mscorlib.dll is hardwired

Also, only build one copy of proto (the one for  .NET 4.0) to reduce
build times