Use mono-sgen for /resident server if mono-sgen is being used for the client
The /resident server is being spawned with mono even if mono-when is
being used for the calling fsc.exe. This change makes the /resident
switch also benefit from SGEN when it is being used.

Will ask on the mono list if there is a better way of detecting if
mono-sgen is being used (or even just finding the exact path to the
mono being used to run the program)