Towards bootstrapping the compiler with xbuild
Wouldn't it be good if we could build the compiler with xbuild (at
least be able to on Mono 3.0.0). This moves in this direction.

The "mono20" and "mono40" target frameworks are removed since we just
use "net20" and "net40" to build.

The FSharpSource.targets is adjusted to use lib/bootstrap/4.0 and to
always use the msft.pubkey (as we do with the Makefile build)

The CustomCopyLocalFiles and subst.exe stuff is removed to simplify the
build

The FSharp.SrGen.targets is copied from the LKG directory and adjusted
to give the right names to the resource files (don't put the outputs in
a target directory)