Fixes to allow optional bootstrap using xbuild and loading compiler .fsproj into MonoDevelop
We can now bootstrap the compiler with xbuild, though it is not the
default

The compiler projec files can also be loaded into MonoDevelop though it
is slow and blocks the UI often (too much garbage being produced? is
MonoDevelop using SGEN?). Added an all-vs2010.sln which can be loaded
into MD

This removes .NET 2.0 as a target for the compiler binaries whcih makes
everything simpler (the library binaries for .NET 2.0 are still
compiled)