Use strong names for bootstrap compiler to help with compile on windows
switch to strong names for compiler binaries and update the bootstrap
compiler with those names. The strong names are based on
src\fsharp\test.snk

this means the strong name for FSharp.Compiler.dll changes, but no one
should be depending on that and the new names are useful as the won't
collide with the visual studio versions of these DLLs on windows. The
monodevelop binding is neutral to the strong names used.

we still compile FSharp.Core as
delay-signed-with-the-microsoft-key-and-then-mono-key-signed which Mono
recognizes as strong named but i think windows does not. We have to use
this identity for FSHarp.Core to keep binary compatibility with
Windows-compiled thing like type provider DLLs.

We may have to use the Microsoft FSHarp.Core.dll on Windows, i think it
gets picked up automatically from the GAC but need to check tomorrow.

also fixes bugs in ilsupp.fs in strong nae signing.

tested by bootstrapping, and manually examining the names on the dlls.