Make mono/xbuild F# resource naming a bit more consistent with Visual F#
See https://github.com/fsharp/fsharp/issues/50

The comments in FSharpBuild/Fsc.fs explain a bit more. It is realy hard
to work out how to make this perfectly consistent. Essentially, be
careful using resources in folders if you want your projects to be
portable across VisualStudio/MSBuild and mono/MonoDevelop/XBuild

We hit this bug in fsharp/fsharpbinding, but we've moved that to no
longer put resources in folders for now.