Enable build of FSharp.Data.TypeProviders.dll
This enables the build of FSharp.Data.TypeProviders.dll via xbuild. The
WSDL functionality is disabled for nowdue to
ConfigurationChannelFactory not being defined on Mono.

None of the functionality is likely to work on Linux or Mac because the
command-line tools used by the type providers will not be found by the
registry searches used on Windows.