install: stop using current mono prefix to install, respect --prefix arg
Even if the $prefix arg seems completely wrong (as in, there is no mono
installation in that prefix):
- The scripts should respect it (because creating .deb packages involves
installing in sandboxed filesystem paths to extract only the relevant bits
installed by a tarball).
- Anyway there is already a warning printed by the configure.ac, just in
case who is compiling is a person instead of a buildbot.
- When compiling, proper mono libs of the existing current installation
will still be properly referenced, so building will still work.

This is done by using $gacdir variables instead of the $monogacdir ones
for the install Makefile targets (which reference /usr/local/lib/mono
instead of /usr/lib/mono, for the case of having mono installed in /usr
but choosing the /usr/local prefix).

This commit finishes DESTDIR support, to work outside of the box, for example
with a .deb package creator like fpm (see a sample script at
https://github.com/knocte/7digital-mono/blob/master/build-fsharp-package.bash ).