install: avoid using DESTDIR in .Targets files
$DESTDIR is only meant to be a prefix to the prefix specified
by the user, at install time. That is, if the prefix specified
was /usr, and $DESTDIR is assigned the value /tmp when running
the target 'make install', then all the files installed should
be in /tmp/usr, but their contents should never reference this
temporal path /tmp/usr, but /usr, because DESTDIR is only
meant to be used as a hook to where leave the binaries generated
for a certain prefix.

Fixes github issue #204, reported by Henrik Feldt.