Make fsi aware of what it is called
On mono fsi/fsc are called through fsharpi/fsharpc using `exec`.  As a
result, fsi doesn't know it is really called fsharpi.  This produced a
weird error message in `#help`.  This patch adds an `exename` parameter
to specify the command that was actually input to run fsi, making the
error messages make sense.

This is for github issue #227