Readme: slightly simplify & clarify FSharp.Core NuGet section
Previously we used multiple ways to refer to the packages which was quite
confusing, on top of the fact that the package naming and versioning is
very confusing in itself already. I suggest we simplify and clarify the
readme section by using only one human readable title which is close or
identical to the NuGet package title, and the NuGet package id itself.

It seems the nuspec files better be renamed to match their id as well,
but I did not touch them to avoid breaking anything.