Fix mono 4.0 installation.
Mono 4.0 no longer ships a functioning 4.0 profile. $install/lib/mono/4.0 only has reference assemblies with no IL.

Which means we cannot have fsc.exe there as mono will try to use the mscorlib on that directory and fail.

The solution is to change the build to default to install things on 4.5 and symlink the libraries on 4.0.